everything

+++ src/main/java/com/el/ally/invoicify/Repositories/CompanyRepository.java
@@ -7,7 +7,7 @@ import com.el.ally.invoicify.models.Company;
 
 public interface CompanyRepository extends JpaRepository<Company, Integer> {
 
-	Company findOne(long clientId);
+	Company findOne(int clientId);
 
 }
